Safari-611.1.10/LayoutTests/media/video-ended-seek-crash-expected.txt


EVENT(canplaythrough)
RUN(video.currentTime = video.duration - 0.1)
RUN(video.play())
EVENT(ended)
RUN(video.setAttribute("loop", "loop"))
RUN(video.currentTime = 0)
RUN(video.play())
EVENT(seeked)
END OF TEST
